# PERF: improve perf of writing csv's with datetimes