```

-------------------------------------------------------------------------------
Test name                                    | head[ms] | base[ms] |  ratio   |
-------------------------------------------------------------------------------
frame_to_csv_mixed                           | 107.7774 | 668.5090 |   0.1612 |
frame_to_csv                                 |  55.9403 | 140.1517 |   0.3991 |
write_csv_standard                           |  25.6400 |  49.6367 |   0.5166 |
packers_write_csv                            | 663.6736 | 1228.0974 |   0.5404 |
read_csv_standard                            |  10.2447 |  15.0284 |   0.6817 |
read_csv_skiprows                            |  12.1686 |  16.3487 |   0.7443 |
frame_to_csv_date_formatting                 |  10.2980 |  13.3820 |   0.7695 |
```